extract validateDate to helpers file
