+++ src/helpers.js
@@ -6,3 +6,11 @@ export const fetchData = async urlParams => {
   const data = await response.json();
   return data;
 };
+
+export const validateDate = (valData, value) => {
+  const { startDate, endDate } = valData;
+  const start = Date.parse(startDate);
+  const end = Date.parse(endDate);
+  const date = Date.parse(value);
+  return date >= start && date <= end && value.length === 10;
+};
